fix question and answer option

+++ askit_frontend/src/views/QuestionView.vue
@@ -37,19 +37,19 @@
                             >
                                 <div class="py-1">
 
-                                    <button
-                                            class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 hover:text-gray-900"
-                                            @click="editItem">
-                                        <svg style="color: blue" xmlns="http://www.w3.org/2000/svg" width="20"
-                                             height="20"
-                                             fill="currentColor" class="mr-3 bi bi-pencil" viewBox="0 0 16 16">
-                                            <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0
-          1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0
-          1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z"
-                                                  fill="blue"></path>
-                                        </svg>
-                                        Edit
-                                    </button>
+<!--                                    <button-->
+<!--                                            class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 hover:text-gray-900"-->
+<!--                                            @click="editItem">-->
+<!--                                        <svg style="color: blue" xmlns="http://www.w3.org/2000/svg" width="20"-->
+<!--                                             height="20"-->
+<!--                                             fill="currentColor" class="mr-3 bi bi-pencil" viewBox="0 0 16 16">-->
+<!--                                            <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0-->
+<!--          1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0-->
+<!--          1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z"-->
+<!--                                                  fill="blue"></path>-->
+<!--                                        </svg>-->
+<!--                                        Edit-->
+<!--                                    </button>-->
 
                                     <button v-on:click="deleteQuestion"
                                             class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-red-700 hover:bg-gray-100 hover:text-red-900">
@@ -61,18 +61,18 @@
                                         Delete
                                     </button>
 
-                                    <button v-on:click="muteNotifications"
-                                            class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-black hover:bg-gray-100 hover:text-black">
-                                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20"
-                                             fill="currentColor"
-                                             class="mr-3 bi bi-bell" viewBox="0 0 16 16">
-                                            <path d="M8 16a2 2 0 0 0 2-2H6a2 2 0 0 0 2 2zM8 1.918l-.797.161A4.002 4.002 0 0 0 4 6c0 .628-.134 2.197-.459 3.742-.16.767-.376
-            1.566-.663 2.258h10.244c-.287-.692-.502-1.49-.663-2.258C12.134 8.197 12 6.628 12 6a4.002 4.002 0 0 0-3.203-3.92L8 1.917zM14.22
-            12c.223.447.481.801.78 1H1c.299-.199.557-.553.78-1C2.68 10.2 3 6.88 3 6c0-2.42 1.72-4.44 4.005-4.901a1 1 0 1 1 1.99 0A5.002 5.002 0 0
-            1 13 6c0 .88.32 4.2 1.22 6z"/>
-                                        </svg>
-                                        Mute Notifications
-                                    </button>
+<!--                                    <button v-on:click="muteNotifications"-->
+<!--                                            class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-black hover:bg-gray-100 hover:text-black">-->
+<!--                                        <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20"-->
+<!--                                             fill="currentColor"-->
+<!--                                             class="mr-3 bi bi-bell" viewBox="0 0 16 16">-->
+<!--                                            <path d="M8 16a2 2 0 0 0 2-2H6a2 2 0 0 0 2 2zM8 1.918l-.797.161A4.002 4.002 0 0 0 4 6c0 .628-.134 2.197-.459 3.742-.16.767-.376-->
+<!--            1.566-.663 2.258h10.244c-.287-.692-.502-1.49-.663-2.258C12.134 8.197 12 6.628 12 6a4.002 4.002 0 0 0-3.203-3.92L8 1.917zM14.22-->
+<!--            12c.223.447.481.801.78 1H1c.299-.199.557-.553.78-1C2.68 10.2 3 6.88 3 6c0-2.42 1.72-4.44 4.005-4.901a1 1 0 1 1 1.99 0A5.002 5.002 0 0-->
+<!--            1 13 6c0 .88.32 4.2 1.22 6z"/>-->
+<!--                                        </svg>-->
+<!--                                        Mute Notifications-->
+<!--                                    </button>-->
                                 </div>
                             </div>
                         </div>
@@ -266,30 +266,30 @@
                                     <button v-if="user.username === answer.author"
                                             class="transition  focus:outline-none mt-3 mr-2 mb-2 rounded-lg bg-blue-600 px-4 text-sm font-medium text-white py-2 hover:bg-blue-500 focus:ring-4 focus:ring-pink-400"
                                             type="button"
-                                            @click="toggleAnswerDropdown">
+                                            @click="toggleAnswerDropdown(answer)">
                                         <span>Options</span>
                                     </button>
 
                                     <div
                                             class="origin-top-right absolute right-0 mt-2 w-56 rounded-md shadow-lg bg-white ring-1 ring-black ring-opacity-5 divide-y divide-gray-100"
-                                            v-show="showDropdownAnswers">
+                                            v-show="answer.showDropdown">
                                         <div class="py-1">
 
-                                            <button @click="editItem"
-                                                    class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 hover:text-gray-900"
-                                            >
-                                                <svg style="color: blue" xmlns="http://www.w3.org/2000/svg"
-                                                     width="20"
-                                                     height="20"
-                                                     fill="currentColor" class="mr-3 bi bi-pencil"
-                                                     viewBox="0 0 16 16">
-                                                    <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0
-                    1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0
-                    1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z"
-                                                          fill="blue"></path>
-                                                </svg>
-                                                Edit Answer
-                                            </button>
+<!--                                            <button @click="editItem"-->
+<!--                                                    class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-gray-700 hover:bg-gray-100 hover:text-gray-900"-->
+<!--                                            >-->
+<!--                                                <svg style="color: blue" xmlns="http://www.w3.org/2000/svg"-->
+<!--                                                     width="20"-->
+<!--                                                     height="20"-->
+<!--                                                     fill="currentColor" class="mr-3 bi bi-pencil"-->
+<!--                                                     viewBox="0 0 16 16">-->
+<!--                                                    <path d="M12.146.146a.5.5 0 0 1 .708 0l3 3a.5.5 0 0 1 0 .708l-10 10a.5.5 0 0 1-.168.11l-5 2a.5.5 0 0 1-.65-.65l2-5a.5.5 0 0-->
+<!--                    1 .11-.168l10-10zM11.207 2.5 13.5 4.793 14.793 3.5 12.5 1.207 11.207 2.5zm1.586 3L10.5 3.207 4 9.707V10h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0-->
+<!--                    1 .5.5v.5h.293l6.5-6.5zm-9.761 5.175-.106.106-1.528 3.821 3.821-1.528.106-.106A.5.5 0 0 1 5 12.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.468-.325z"-->
+<!--                                                          fill="blue"></path>-->
+<!--                                                </svg>-->
+<!--                                                Edit Answer-->
+<!--                                            </button>-->
 
                                             <button v-on:click="deleteAnswer(answer.id)"
                                                     class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-red-700 hover:bg-gray-100 hover:text-red-900">
@@ -303,18 +303,18 @@
                                                 Delete Answer
                                             </button>
 
-                                            <button v-on:click="muteNotifications"
-                                                    class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-black hover:bg-gray-100 hover:text-black">
-                                                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20"
-                                                     fill="currentColor"
-                                                     class="mr-3 bi bi-bell" viewBox="0 0 16 16">
-                                                    <path d="M8 16a2 2 0 0 0 2-2H6a2 2 0 0 0 2 2zM8 1.918l-.797.161A4.002 4.002 0 0 0 4 6c0 .628-.134 2.197-.459 3.742-.16.767-.376
-                      1.566-.663 2.258h10.244c-.287-.692-.502-1.49-.663-2.258C12.134 8.197 12 6.628 12 6a4.002 4.002 0 0 0-3.203-3.92L8 1.917zM14.22
-                      12c.223.447.481.801.78 1H1c.299-.199.557-.553.78-1C2.68 10.2 3 6.88 3 6c0-2.42 1.72-4.44 4.005-4.901a1 1 0 1 1 1.99 0A5.002 5.002 0 0
-                      1 13 6c0 .88.32 4.2 1.22 6z"/>
-                                                </svg>
-                                                Mute Notifications
-                                            </button>
+<!--                                            <button v-on:click="muteNotifications"-->
+<!--                                                    class="transition focus:ring-4 focus:outline-none focus:ring-blue-400 flex justify-start w-full px-4 py-2 text-sm text-black hover:bg-gray-100 hover:text-black">-->
+<!--                                                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20"-->
+<!--                                                     fill="currentColor"-->
+<!--                                                     class="mr-3 bi bi-bell" viewBox="0 0 16 16">-->
+<!--                                                    <path d="M8 16a2 2 0 0 0 2-2H6a2 2 0 0 0 2 2zM8 1.918l-.797.161A4.002 4.002 0 0 0 4 6c0 .628-.134 2.197-.459 3.742-.16.767-.376-->
+<!--                      1.566-.663 2.258h10.244c-.287-.692-.502-1.49-.663-2.258C12.134 8.197 12 6.628 12 6a4.002 4.002 0 0 0-3.203-3.92L8 1.917zM14.22-->
+<!--                      12c.223.447.481.801.78 1H1c.299-.199.557-.553.78-1C2.68 10.2 3 6.88 3 6c0-2.42 1.72-4.44 4.005-4.901a1 1 0 1 1 1.99 0A5.002 5.002 0 0-->
+<!--                      1 13 6c0 .88.32 4.2 1.22 6z"/>-->
+<!--                                                </svg>-->
+<!--                                                Mute Notifications-->
+<!--                                            </button>-->
                                         </div>
                                     </div>
                                     <div v-if="user.username === question.author"
@@ -429,8 +429,8 @@ export default {
             this.showDropdown = !this.showDropdown;
         },
 
-        toggleAnswerDropdown() {
-            this.showDropdownAnswers = !this.showDropdownAnswers;
+        toggleAnswerDropdown(answer) {
+            answer.showDropdown = !answer.showDropdown;
         },
         addAnswer() {
             console.log(this.answerInput)
@@ -477,6 +477,9 @@ export default {
         },
 
         deleteAnswer(answerId) {
+            if (!confirm("Are you sure you want to delete this answer?")) {
+                return;
+            }
             axiosClient.delete(`/question/${this.$route.params.id}/answer/${answerId}/delete_answer/`)
                 .then((response) => {
                     window.location.href = `/question/${this.$route.params.id}`;
@@ -547,6 +550,9 @@ export default {
         },
 
         deleteQuestion() {
+            if (window.confirm("Are you sure you want to delete this question?") === false) {
+                return;
+            }
             axiosClient.delete(`/question/${this.$route.params.id}/delete/`)
                 .then((response) => {
                     window.location.href = `/module/${response.data.module}`;
